feat: added client kafka fix #44

+++ finrv.Domain/Business/AssetPriceAverager/Models/TradeData.cs
@@ -3,7 +3,7 @@ using finrv.Domain.Enums;
 namespace finrv.Domain.Business.AssetPriceAverager.Models;
 
 public record TradeData(
-    long PositionSize,
+    decimal PositionSize,
     ETransactionType Type,
     decimal Brokerage,
     decimal Price,
